If no (or invalid) articles, then render nothing.
